Complete game func
Added for loop to game func in order for it to play 5 rounds and log
the results of each round to the console. Added if/else statements
in order to log the winner of the overall game. Added variable so that
the winner message for a round could be logged.