Add style modify index and script
Added buttons in a list in index in order to prepare for proper logic
in script file for playing the game. Added id's to style file in order
to link them later with their respective buttons. Added class for
styling in the future. Added styling prep for list of buttons. Removed
playing 5 games logic from script as per the prompt.