Defer script and add button logic
Added defer tag to script so the page would load first and then run the
logic. Added logic to script file to output the correct player choice
based on the button pressed. Removed the string checking logic in the
playRound function because the prompt no longer asks for a string input