Add text feedback functionality
Added a new div with id in order to hold the text that will be
displayed every round. Added id to css file for future use. Added DOM
methods to script file in order to display text showing who won and to
overwrite that text with the winner every round.